+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.0
+
+### BazelProcess
+
+- Added `String directory: ...` (CWD) as an optional argument to constructor.
+- Added `runBazel`.
+- Changed `query` to take a `List<String> options`.
+
 ## 0.0.0
 
 - Initial commit.